move tokenizer to js folder
